requirements.txt
kubernetes
asyncio 
google.adk 
pika 